backend/requirements.txt
# ===============================
# FastAPI & Server
# ===============================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===============================
# Database
# ===============================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ===============================
# Authentication
# ===============================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===============================
# Core ML / Math
# ===============================
numpy<2.0.0
scipy==1.11.4
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2

# ===============================
# Deep Learning
# ===============================
torch==2.1.0

# ===============================
# GNN (ABSOLUTELY REQUIRED)
# ===============================
torch-geometric==2.4.0
torch-scatter
torch-sparse
networkx==3.2.1

# ===============================
# Explainability
# ===============================
shap==0.43.0

# ===============================
# Realtime / Async
# ===============================
websockets==12.0
celery==5.3.4
redis==5.0.1

# ===============================
# Utilities / Testing
# ===============================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
fpdf2==2.7.7
